+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "cph-ng" extension will be documented in this file.
 
+## [Unreleased]
+
+- **fix**: the cache telemetry log is too much
+
 ## 0.4.12
 
 - **feat**: integrate telemetry for performance tracking and analytics